requirements.txt
awscli==1.9.5
backports.ssl-match-hostname==3.4.0.2
BigQuery-Python==1.4.0
botocore==1.3.5
certifi==2015.9.6.2
cffi==1.3.0
colorama==0.3.3
cryptography==1.1
docutils==0.12
enum34==1.0.4
google-api-python-client==1.4.2
httplib2==0.9.2
idna==2.0
ipaddress==1.0.14
jmespath==0.9.0
oauth2client==1.5.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pyOpenSSL==0.15.1
python-dateutil==2.4.2
PyYAML==3.11
rsa==3.2.3
simplejson==3.8.1
six==1.10.0
tornado==4.2.1
uritemplate==0.6
wheel==0.24.0
